+++ Samples/SetappSample-macOS-ObjectiveC/ViewController.h
@@ -1,0 +1,14 @@
+//
+//  ViewController.h
+//  SetappSample-macOS-ObjectiveC
+//
+//  Created by Aleksandr.Bilous on 22.04.2022.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface ViewController : NSViewController
+
+
+@end
+