+++ Samples/SetappSample-macOS-ObjectiveC/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  SetappSample-macOS-ObjectiveC
+//
+//  Created by Aleksandr.Bilous on 22.04.2022.
+//
+
+#import <Cocoa/Cocoa.h>
+
+int main(int argc, const char * argv[])
+{
+    @autoreleasepool {
+        // Setup code that might create autoreleased objects goes here.
+    }
+    return NSApplicationMain(argc, argv);
+}